Rename components/DataSelector.vue to components/index/_shared/DataSelector.vue
